New TweetFetcher (still not finished), will try to fetch the newest pages of #DiccionarioChimbo Tweets.
New Settings class. A Class to store application wide settings.
Settings are stored in MongoDB. All you need to know about this class
is .instance(), .load() and .save().

Updated Tweet to make ids long. Updated the MongoMapper to map the
Settings class.

+++ src/indexer/java/com/nolapeles/diccionariochimbo/indexer/models/Tweet.java
@@ -17,7 +17,7 @@ public class Tweet {
 	public ObjectId id;
 	
 	@Indexed
-	public int tweet_id;
+	public long tweet_id;
 	
 	public String text;
